Moved roles to separate component
